[PVR] Refactor CPVRGUIActions step 3: Move EPG related actions to own class.
