+++ xbmc/pvr/guilib/CMakeLists.txt
@@ -2,6 +2,7 @@ set(SOURCES GUIEPGGridContainer.cpp
             GUIEPGGridContainerModel.cpp
             PVRGUIActionListener.cpp
             PVRGUIActions.cpp
+            PVRGUIActionsEPG.cpp
             PVRGUIActionsRecordings.cpp
             PVRGUIActionsTimers.cpp
             PVRGUIChannelIconUpdater.cpp
@@ -12,6 +13,7 @@ set(HEADERS GUIEPGGridContainer.h
             GUIEPGGridContainerModel.h
             PVRGUIActionListener.h
             PVRGUIActions.h
+            PVRGUIActionsEPG.h
             PVRGUIActionsRecordings.h
             PVRGUIActionsTimers.h
             PVRGUIChannelIconUpdater.h
