+++ xbmc/pvr/guilib/PVRGUIActions.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "pvr/PVRChannelNumberInputHandler.h"
+#include "pvr/guilib/PVRGUIActionsEPG.h"
 #include "pvr/guilib/PVRGUIActionsRecordings.h"
 #include "pvr/guilib/PVRGUIActionsTimers.h"
 #include "pvr/guilib/PVRGUIChannelNavigator.h"
@@ -65,32 +66,13 @@ namespace PVR
     void SwitchToPreviousChannel();
   };
 
-  class CPVRGUIActions : public CPVRGUIActionsRecordings, public CPVRGUIActionsTimers
+  class CPVRGUIActions : public CPVRGUIActionsEPG,
+                         public CPVRGUIActionsRecordings,
+                         public CPVRGUIActionsTimers
   {
   public:
     CPVRGUIActions();
     virtual ~CPVRGUIActions() = default;
-
-    /*!
-     * @brief Open a dialog with epg information for a given item.
-     * @param item containing epg data to show. item must be an epg tag, a channel or a timer.
-     * @return true on success, false otherwise.
-     */
-    bool ShowEPGInfo(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open a dialog with the epg list for a given item.
-     * @param item containing channel info. item must be an epg tag, a channel or a timer.
-     * @return true on success, false otherwise.
-     */
-    bool ShowChannelEPG(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open a window containing a list of epg tags 'similar' to a given item.
-     * @param item containing epg data for matching. item must be an epg tag, a channel or a recording.
-     * @return true on success, false otherwise.
-     */
-    bool FindSimilar(const std::shared_ptr<CFileItem>& item) const;
 
     /*!
      * @brief Get a localized resume play label, if the given item can be resumed.
@@ -287,34 +269,6 @@ namespace PVR
      * @param item The item.
      */
     bool OnInfo(const std::shared_ptr<CFileItem>& item);
-
-    /*!
-     * @brief Execute a saved search. Displays result in search window if it is open.
-     * @param item The item containing a search filter.
-     * @return True on success, false otherwise.
-     */
-    bool ExecuteSavedSearch(const std::shared_ptr<CFileItem>& item);
-
-    /*!
-     * @brief Edit a saved search. Opens the search dialog.
-     * @param item The item containing a search filter.
-     * @return True on success, false otherwise.
-     */
-    bool EditSavedSearch(const std::shared_ptr<CFileItem>& item);
-
-    /*!
-     * @brief Rename a saved search. Opens a title input dialog.
-     * @param item The item containing a search filter.
-     * @return True on success, false otherwise.
-     */
-    bool RenameSavedSearch(const std::shared_ptr<CFileItem>& item);
-
-    /*!
-     * @brief Delete a saved search. Opens confirmation dialog before deleting.
-     * @param item The item containing a search filter.
-     * @return True on success, false otherwise.
-     */
-    bool DeleteSavedSearch(const std::shared_ptr<CFileItem>& item);
 
   private:
     CPVRGUIActions(const CPVRGUIActions&) = delete;
